fix: verify if already exists the line before adding

+++ hooks/android/build_gradle_add_dependency.js
@@ -6,14 +6,26 @@ module.exports = function(ctx) {
     readline = require("readline"),
     deferral = require('q').defer();
 
+var googleServicesStr = "if (project.extensions.findByName('googleServices') == null) { apply plugin: 'com.google.gms.google-services' }"
+var googleServicesStrExists = false
+var classpathsStrToVerify = "com.google.gms:google-services:4.3.10"
+var classpathsStr = '\t\tclasspath "com.google.gms:google-services:4.3.10"'
+var classpathsExists = false
+
 var lineReader = readline.createInterface({
     terminal: false,
     input : fs.createReadStream('platforms/android/build.gradle')
 });
 lineReader.on("line", function(line) {
+    if (line.includes(classpathsStrToVerify)) {
+        classpathsExists = true
+    }
+
     fs.appendFileSync('./build.gradle', line.toString() + os.EOL);
     if (/.*\ dependencies \{.*/.test(line)) {
-        fs.appendFileSync('./build.gradle', '\t\tclasspath "com.google.gms:google-services:4.3.10"' + os.EOL);
+        if (!classpathsExists) {
+            fs.appendFileSync('./build.gradle', classpathsStr + os.EOL);
+        }
     }
 }).on("close", function () {
     fs.rename('./build.gradle', 'platforms/android/build.gradle', deferral.resolve);
@@ -23,10 +35,16 @@ var lineReaderApp = readline.createInterface({
     terminal: false,
     input : fs.createReadStream('platforms/android/app/build.gradle')
 });
+lineReaderApp.on("line", function (line) {
+    if (line.includes(googleServicesStr)) {
+        googleServicesStrExists = true
+    }
+});    
 lineReaderApp.on("close", function () {
-    var lineToAdd = "if (project.extensions.findByName('googleServices') == null) { apply plugin: 'com.google.gms.google-services' }"
-    fs.appendFileSync('./platforms/android/app/build.gradle', lineToAdd + os.EOL);
-    fs.rename('./platforms/android/app/build.gradle', 'platforms/android/app/build.gradle', deferral.resolve);
+    if (!googleServicesStrExists) {
+        fs.appendFileSync('./platforms/android/app/build.gradle', googleServicesStr + os.EOL);
+        fs.rename('./platforms/android/app/build.gradle', 'platforms/android/app/build.gradle', deferral.resolve);    
+    }
 });
 
 return deferral.promise;
